refactor: make widgets for size and color controls
