+++ glue_jupyter/__init__.py
@@ -1,19 +1,7 @@
 from __future__ import absolute_import
 from glue.core import message as msg
-from glue.core.subset import Subset
-from glue.viewers.common.viewer import Viewer
-from glue.core.layer_artist import LayerArtistContainer
-from glue.core.edit_subset_mode import (EditSubsetMode, OrMode, AndNotMode,
-                                        AndMode, XorMode, ReplaceMode)
-from glue.core.roi import PolygonalROI, CircularROI, RectangularROI, Projected3dROI
-from glue.core.subset import RoiSubsetState3d, RoiSubsetState
-from glue.core.command import ApplySubsetState
-from IPython.display import display
-from glue.core.link_helpers import LinkSame
-import ipywidgets as widgets
 import six
 
-from glue_jupyter.utils import _update_not_none
 # from glue.core.session import Session
 # from glue.viewers.scatter.layer_artist import ScatterLayerArtist
 
@@ -27,6 +15,7 @@ def jglue(*args, **kwargs):
     from glue.app.qt import GlueApplication
     from glue.qglue import parse_data, parse_links
     from glue.core.data_factories import load_data
+    from .app import JupyterApplication
 
     links = kwargs.pop('links', None)
 
@@ -80,409 +69,5 @@ def example_image(shape=64, limits=[-4, 4]):
     data.add_component(I, label='intensity')
     return data
 
-# not sure we need to inherit: from glue.core.application_base import Application
-# what would we gain that would be natural in the notebook?
-from glue.core.application_base import Application
 
-class JupyterApplication(Application):
 
-    def __init__(self, data_collection=None, session=None):
-        super(JupyterApplication, self).__init__(data_collection=data_collection, session=session)
-        self.selection_modes = [('replace', ReplaceMode), ('add', OrMode), ('and', AndMode), ('xor', XorMode), ('remove', AndNotMode)]
-        self.widget_selection_mode = widgets.ToggleButtons(
-            options=[label for label, mode in self.selection_modes],
-            description='Selection mode:',
-            disabled=False,
-            tooltips=[label for label, mode in self.selection_modes],
-        )
-        self.widget_data_collection = widgets.SelectMultiple()
-        self.widget_subset_groups   = widgets.SelectMultiple()
-        self.widget = widgets.VBox(children=[self.widget_selection_mode, self.widget_subset_groups])
-        self.widget_selection_mode.observe(self._set_selection_mode, 'index')
-        self.widget_subset_groups.observe(self._set_subset_groups, 'index')
-        self.session.hub.subscribe(self, msg.EditSubsetMessage, handler=self._on_edit_subset_msg)
-        self.session.hub.subscribe(self, msg.SubsetCreateMessage, handler=self._on_subset_create_msg)
-        self._update_subset_mode(self.session.edit_subset_mode.mode)
-        self._update_subset_groups_selected(self.session.edit_subset_mode.edit_subset)
-        display(self.widget)
-
-    def link(self, links):
-        from glue.qglue import parse_links
-        self.data_collection.add_link(parse_links(self.data_collection, links))
-
-    def add_link(self, data1, attribute1, data2, attribute2, function=None):
-        # For now this assumes attribute1 and attribute2 are strings and single
-        # attributes. In future we should generalize this while keeping the
-        # simplest use case simple.
-        if function is not None:
-            raise NotImplementedError
-        att1 = data1.id[attribute1]
-        att2 = data2.id[attribute2]
-        link = LinkSame(att1, att2)
-        self.data_collection.add_link(link)
-
-    def _on_edit_subset_msg(self, msg):
-        self._update_subset_mode(msg.mode)
-        self._update_subset_groups_selected(msg.subset)
-
-    def _on_subset_create_msg(self, msg):
-        self._update_subset_groups_selected(self.session.edit_subset_mode.edit_subset)
-
-    def _update_subset_mode(self, mode):
-        if self.session.edit_subset_mode.mode != mode:
-            self.session.edit_subset_mode.mode = mode
-        index = 0
-        EditSubsetMode
-        for i, (name, sel_mode) in enumerate(self.selection_modes):
-            if mode == sel_mode:
-                index = i
-        self.widget_selection_mode.index = index
-
-    def _update_subset_groups_selected(self, subset_groups_selected):
-        if self.session.edit_subset_mode.edit_subset != subset_groups_selected:
-            self.session.edit_subset_mode.edit_subset = subset_groups_selected
-        options = []
-        indices = []
-        for i, subset_group in enumerate(self.data_collection.subset_groups):
-            options.append(subset_group.label)
-            if subset_group in self.session.edit_subset_mode.edit_subset:
-                indices.append(i)
-        self.widget_subset_groups.options = options
-        self.widget_subset_groups.index = tuple(indices)
-
-    def _set_subset_groups(self, change):
-        subset_groups = [self.data_collection.subset_groups[k] for k in self.widget_subset_groups.index]
-        self.session.edit_subset_mode.edit_subset = subset_groups
-
-    def _set_selection_mode(self, change):
-        #EditSubsetMode().mode = self.selection_modes[change.new][1]
-        self.subset_mode(self.selection_modes[change.new][1])
-
-    def subset_mode(self, mode):
-        self.session.edit_subset_mode.mode = mode
-
-    def subset_mode_replace(self):
-        self.subset_mode(ReplaceMode)
-
-    def subset_mode_and(self):
-        self.subset_mode(AndMode)
-
-    def subset_lasso2d(self, x, y, xvalues, yvalues):
-        roi = PolygonalROI(xvalues, yvalues)
-        self.subset_roi([x, y], roi)
-
-    def subset_roi(self, components, roi, use_current=False):
-        subset_state = RoiSubsetState(components[0], components[1], roi)
-        cmd = ApplySubsetState(data_collection=self.data_collection,
-                               subset_state=subset_state,
-                               use_current=use_current)
-        self._session.command_stack.do(cmd)
-
-    def _roi_to_subset_state(self, components, roi):
-        return RoiSubsetState(components[0], components[1], roi)
-
-    def add_widget(self, widget, label=None, tab=None):
-        pass
-
-    # TODO: remove when https://github.com/glue-viz/glue/pull/1877 is merged
-    def new_data_viewer(self, viewer_class, data=None, state=None):
-        """
-        Create a new data viewer, add it to the UI,
-        and populate with data
-
-        """
-        from glue.core import BaseData
-
-        if viewer_class is None:
-            return
-
-        if state is not None:
-            c = viewer_class(self._session, state=state)
-        else:
-            c = viewer_class(self._session)
-        c.register_to_hub(self._session.hub)
-
-        if data is not None:
-            if isinstance(data, BaseData):
-                result = c.add_data(data)
-            elif isinstance(data, Subset):
-                result = c.add_subset(data)
-            if not result:
-                c.close(warn=False)
-                return
-
-        self.add_widget(c)
-        c.show()
-        return c
-
-
-    def histogram1d(self, x=None, data=None, widget='bqplot', color=None, x_min=None, x_max=None, hist_n_bin=None, normalize=False, cumulative=False, viewer_state=None, layer_state=None):
-        if widget == 'bqplot':
-            from .bqplot import BqplotHistogramView
-            viewer_cls = BqplotHistogramView
-        elif widget == 'matplotlib':
-            from .matplotlib.histogram import HistogramJupyterViewer
-            viewer_cls = HistogramJupyterViewer
-        else:
-            raise ValueError("Widget type should be 'bqplot' or 'matplotlib'")
-        if data is None and len(self._data) != 1:
-            raise ValueError('There is more than 1 data set in the data collection, please pass a data argument')
-        data = data or self._data[0]
-
-        state = viewer_cls._state_cls()
-        state.x_att_helper.append_data(data)
-
-        viewer_state_obj = viewer_cls._state_cls()
-        viewer_state_obj.x_att_helper.append_data(data)
-        viewer_state = viewer_state or {}
-        if x is not None:
-            viewer_state['x_att'] = data.id[x]
-        # x_min and x_max get set to the hist_x_min/max in glue.viewers.histogram.state
-        # for this API it make more sense to call it x_min and x_max, and for consistency with the rest
-        _update_not_none(viewer_state, hist_x_min=x_min, hist_x_max=x_max, hist_n_bin=hist_n_bin,
-            normalize=normalize, cumulative=cumulative)
-        viewer_state_obj.update_from_dict(viewer_state)
-
-        view = self.new_data_viewer(viewer_cls, data=data, state=viewer_state_obj)
-        layer_state = layer_state or {}
-        _update_not_none(layer_state, color=color)
-        view.layers[0].state.update_from_dict(layer_state)
-        return view
-
-    def scatter2d(self, x=None, y=None, data=None, widget='bqplot', color=None, size=None, viewer_state=None, layer_state=None):
-        if widget == 'bqplot':
-            from .bqplot import BqplotScatterView
-            viewer_cls = BqplotScatterView
-        elif widget == 'matplotlib':
-            from .matplotlib.scatter import ScatterJupyterViewer
-            viewer_cls = ScatterJupyterViewer
-        else:
-            raise ValueError("Widget type should be 'bqplot' or 'matplotlib'")
-        if data is None and len(self._data) != 1:
-            raise ValueError('There is more than 1 data set in the data collection, please pass a data argument')
-        data = data or self._data[0]
-        viewer_state_obj = viewer_cls._state_cls()
-        viewer_state_obj.x_att_helper.append_data(data)
-        viewer_state_obj.y_att_helper.append_data(data)
-        viewer_state = viewer_state or {}
-        if x is not None:
-            viewer_state['x_att'] = data.id[x]
-        if x is not None:
-            viewer_state['y_att'] = data.id[y]
-        viewer_state_obj.update_from_dict(viewer_state)
-
-        view = self.new_data_viewer(viewer_cls, data=data, state=viewer_state_obj)
-        layer_state = layer_state or {}
-        _update_not_none(layer_state, color=color, size=size)
-        view.layers[0].state.update_from_dict(layer_state)
-        return view
-
-    def scatter3d(self, x=None, y=None, z=None, data=None):
-        from .ipyvolume import IpyvolumeScatterView
-        if data is None and len(self._data) != 1:
-            raise ValueError('There is more than 1 data set in the data collection, please pass a data argument')
-        data = data or self._data[0]
-        view = self.new_data_viewer(IpyvolumeScatterView, data=data)
-        if x is not None:
-            x = data.id[x]
-            view.state.x_att = x
-        if y is not None:
-            y = data.id[y]
-            view.state.y_att = y
-        if z is not None:
-            z = data.id[z]
-            view.state.z_att = z
-        return view
-
-    def imshow(self, x=None, y=None, data=None, widget='bqplot'):
-        if widget == 'bqplot':
-            from .bqplot import BqplotImageView
-            viewer_cls = BqplotImageView
-        elif widget == 'matplotlib':
-            from .matplotlib.image import ImageJupyterViewer
-            viewer_cls = ImageJupyterViewer
-        else:
-            raise ValueError("Widget type should be 'bqplot' or 'matplotlib'")
-        data = data or self._data[0]
-        if data is None and len(self._data) != 1:
-            raise ValueError('There is more than 1 data set in the data collection, please pass a data argument')
-        if len(data.pixel_component_ids) < 2:
-            raise ValueError('There are less than 2 pixel components (not an image?)')
-        view = self.new_data_viewer(viewer_cls, data=data)
-        if x is not None:
-            x = data.id[x]
-            view.state.x_att = x
-        if y is not None:
-            y = data.id[y]
-            view.state.y_att = y
-        return view
-
-    def profile1d(self, x, data=None, widget='matplotlib'):
-        if widget == 'matplotlib':
-            from .matplotlib.profile import ProfileJupyterViewer
-            viewer_cls = ProfileJupyterViewer
-        else:
-            raise ValueError("Widget type should be 'matplotlib'")
-        data = data or self._data[0]
-        view = self.new_data_viewer(viewer_cls, data=data)
-        if x is not None:
-            x = data.id[x]
-            view.state.x_att = x
-        return view
-
-    def volshow(self, x="Pixel Axis 2 [x]", y="Pixel Axis 1 [y]", z="Pixel Axis 0 [z]", data=None):
-        from .ipyvolume import IpyvolumeVolumeView
-        data = data or self._data[0]
-        if data is None and len(self._data) != 1:
-            raise ValueError('There is more than 1 data set in the data collection, please pass a data argument')
-        view = self.new_data_viewer(IpyvolumeVolumeView, data=data)
-        if x is not None:
-            x = data.id[x]
-            view.state.x_att = x
-        if y is not None:
-            y = data.id[y]
-            view.state.y_att = y
-        if z is not None:
-            z = data.id[z]
-            view.state.z_att = z
-        return view
-
-    def subset(self, name, state):
-        return self.data_collection.new_subset_group(name, state)
-
-    def _update_undo_redo_enabled(self, *args):
-        pass  # TODO: if we want a gui for this, we need to update it here
-
-    def _choose_merge(self, *args, **kwargs):
-        # Never suggest automatic merging
-        return None, None
-
-
-class IPyWidgetLayerArtistContainer(LayerArtistContainer):
-
-    def __init__(self):
-        super(IPyWidgetLayerArtistContainer, self).__init__()
-        pass  #print('layer artist created')
-
-
-class IPyWidgetView(Viewer):
-
-    _layer_artist_container_cls = IPyWidgetLayerArtistContainer
-
-    def __init__(self, session, state=None):
-        super(IPyWidgetView, self).__init__(session, state=state)
-
-    # TODO: a lot of this comes from DataViewerWithState
-    def register_to_hub(self, hub):
-        super(IPyWidgetView, self).register_to_hub(hub)
-
-        hub.subscribe(self, msg.SubsetCreateMessage,
-                      handler=self._add_subset,
-                      filter=self._subset_has_data)
-
-        hub.subscribe(self, msg.SubsetUpdateMessage,
-                      handler=self._update_subset,
-                      filter=self._has_data_or_subset)
-
-        hub.subscribe(self, msg.SubsetDeleteMessage,
-                      handler=self._remove_subset,
-                      filter=self._has_data_or_subset)
-
-        hub.subscribe(self, msg.NumericalDataChangedMessage,
-                      handler=self._update_data,
-                      filter=self._has_data_or_subset)
-
-        # hub.subscribe(self, msg.DataCollectionDeleteMessage,
-        #               handler=self._remove_data)
-
-        hub.subscribe(self, msg.ComponentsChangedMessage,
-                      handler=self._update_data,
-                      filter=self._has_data_or_subset)
-
-        # hub.subscribe(self, msg.SettingsChangeMessage,
-        #               self._update_appearance_from_settings,
-        #               filter=self._is_appearance_settings)
-
-
-    def add_data(self, data, color=None, alpha=None, **layer_state):
-
-        # Check if data already exists in viewer
-        if not self.allow_duplicate_data and data in self._layer_artist_container:
-            return None
-
-        if self.large_data_size is not None and data.size >= self.large_data_size:
-            proceed = self.warn('Add large data set?', 'Data set {0:s} has {1:d} points, and '
-                                'may render slowly.'.format(data.label, data.size),
-                                default='Cancel', setting='show_large_data_warning')
-            if not proceed:
-                return None
-
-        if data not in self.session.data_collection:
-            raise IncompatibleDataException("Data not in DataCollection")
-
-        # Create layer artist and add to container
-        layer = self.get_data_layer_artist(data)
-
-        if layer is None:
-            return None
-
-        self._layer_artist_container.append(layer)
-        layer_state = dict(layer_state)
-        _update_not_none(layer_state, color=color, alpha=alpha)
-        layer.update()
-        layer.state.update_from_dict(layer_state)
-
-        # Add existing subsets to viewer
-        for subset in data.subsets:
-            self.add_subset(subset)
-
-        return layer
-
-    def remove_subset(self, subset):
-        if subset in self._layer_artist_container:
-            self._layer_artist_container.pop(subset)
-            self.redraw()
-
-    def _remove_subset(self, message):
-        self.remove_subset(message.subset)
-
-    def _add_subset(self, message):
-        self.add_subset(message.subset)
-
-    def _update_subset(self, message):
-        if message.subset in self._layer_artist_container:
-            for layer_artist in self._layer_artist_container[message.subset]:
-                if isinstance(message, msg.SubsetUpdateMessage) and message.attribute not in ['subset_state']:
-                    pass
-                else:
-                    layer_artist.update()
-            self.redraw()
-
-    def _subset_has_data(self, x):
-        return x.sender.data in self._layer_artist_container.layers
-
-    def _has_data_or_subset(self, x):
-        return x.sender in self._layer_artist_container.layers
-
-    def get_layer_artist(self, cls, layer=None, layer_state=None):
-        return cls(self, self.state, layer=layer, layer_state=layer_state)
-
-    def _update_data(self, message):
-        if message.data in self._layer_artist_container:
-            for layer_artist in self._layer_artist_container:
-                if isinstance(layer_artist.layer, Subset):
-                    if layer_artist.layer.data is message.data:
-                        layer_artist.update()
-                else:
-                    if layer_artist.layer is message.data:
-                        layer_artist.update()
-            self.redraw()
-
-    def _add_layer_tab(self, layer):
-        layer_tab = layer.create_widgets()
-        self.tab.children = self.tab.children + (layer_tab, )
-        if isinstance(layer.layer, Subset):
-            label = '{data_label}:{subset_label}'.format(data_label=layer.layer.data.label, subset_label=layer.layer.label)
-        else:
-            label = layer.layer.label
-        self.tab.set_title(len(self.tab.children)-1, label)
