+++ glue_jupyter/ipyvolume/tests/test_ipyvolume.py
@@ -67,6 +67,31 @@ def test_scatter3d(app, dataxyz, dataxz):
     layer.state.vector_visible = True
     assert layer.quiver.visible
 
+
+def test_scatter3d_cmap_mode(app,dataxyz):
+    s = app.scatter3d('x', 'y', data=dataxyz)
+    l1 = s.layers[0]
+    assert l1.state.cmap_mode == 'Fixed', 'expected default value'
+    assert l1.state.cmap_name == 'Gray'
+
+    assert l1.scatter.color.shape == (), 'numpy scalar'
+    l1.state.cmap_att = 'x'
+    l1.state.cmap_mode = 'Linear'
+    assert l1.widget_color.widget_cmap_mode.label == 'Linear'
+    assert l1.state.cmap_name == 'Gray'
+    l1.state.cmap_vmin  = 0
+    l1.state.cmap_vmax  = 10
+    assert l1.scatter.color.shape == (3, 4)
+    assert l1.scatter.color is not None
+
+    l1.widget_color.widget_cmap.label = 'Viridis'
+    assert l1.state.cmap_name == 'Viridis'
+    assert l1.widget_color.widget_cmap.label == 'Viridis'
+
+    l1.widget_color.widget_cmap.label = 'Gray'
+    assert l1.widget_color.widget_cmap.label == 'Gray'
+    assert l1.state.cmap_name == 'Gray'
+
 def test_roi3d(dataxyz):
     roi_2d = PolygonalROI(vx=[0.5, 2.5, 2.5, 0.5], vy=[1, 1, 3.5, 3.5])
     roi = Projected3dROI(roi_2d, projection_matrix=np.eye(4))
