+++ glue_jupyter/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .color import Color
+from .size import Size
